Refactor into analysis package and GUI app
