+++ analysis/__init__.py
@@ -1,0 +1,10 @@
+from .scatter import ScatterAnalysis
+from .histogram import HistogramAnalysis
+from .box import BoxAnalysis
+from .contour import ContourAnalysis
+from .kmeans import KMeansAnalysis
+from .dfa import DFAAnalysis
+from .lstm import LSTMAnalysis
+from .psa import PSAAnalysis
+from .prophet import ProphetAnalysis
+from .mass import MassEstimation
